chore(origin-ui-core): permissions as separate component
