+++ packages/origin-ui-core/src/components/PermissionsFeedback.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import { IDevicePermission, useTranslation } from '../../utils';
+import { useTranslation, IPermission } from '../utils';
 import {
     Typography,
     List,
@@ -10,11 +10,11 @@ import {
 } from '@material-ui/core';
 
 interface IProps {
-    canCreateDevice: IDevicePermission;
+    canAccessPage: IPermission;
 }
 
-export function DevicePermissionsFeedback(props: IProps) {
-    const { canCreateDevice } = props;
+export function PermissionsFeedback(props: IProps) {
+    const { canAccessPage } = props;
     const { t } = useTranslation();
 
     return (
@@ -23,7 +23,7 @@ export function DevicePermissionsFeedback(props: IProps) {
                 {t('info.needToFulfilCriteria')}
             </Typography>
             <List>
-                {canCreateDevice?.rules.map((rule) => (
+                {canAccessPage?.rules.map((rule) => (
                     <ListItem key={rule.label} role={undefined} dense>
                         <ListItemIcon>
                             <Checkbox
