+++ packages/origin-ui-core/src/utils/device.ts
@@ -2,10 +2,6 @@
 import { ProducingDevice } from '@energyweb/device-registry';
 import { IEnvironment } from '../features/general/actions';
 import { CustomFilterInputType, ICustomFilterDefinition } from '../components/Table/FiltersHeader';
-import { getExchangeDepositAddress, getUserOffchain } from '../features/users/selectors';
-import { useSelector } from 'react-redux';
-import { useTranslation } from '.';
-import { OrganizationStatus, UserStatus } from '@energyweb/origin-backend-core';
 import gaseous from '../../assets/device/icon-gaseous.svg';
 import hydro from '../../assets/device/icon-hydro.svg';
 import liquid from '../../assets/device/icon-liquid.svg';
@@ -175,52 +171,6 @@ export function getDeviceSpecificPropertiesSearchTitle(
     }
 
     return t('general.actions.search');
-}
-
-export interface IDevicePermissionRule {
-    label: string;
-    passing: boolean;
-}
-
-export interface IDevicePermission {
-    value: boolean;
-    rules: IDevicePermissionRule[];
-}
-
-export function useDevicePermissions() {
-    const user = useSelector(getUserOffchain);
-    const exchangeAddress = useSelector(getExchangeDepositAddress);
-    const { t } = useTranslation();
-
-    const canCreateDevice: IDevicePermission = {
-        value: false,
-        rules: [
-            {
-                label: t('general.feedback.haveToBeLoggedInUser'),
-                passing: Boolean(user)
-            },
-            {
-                label: t('general.feedback.hasToBeActiveUser'),
-                passing: user?.status === UserStatus.Active
-            },
-            {
-                label: t('general.feedback.userHasToBePartOfApprovedOrganization'),
-                passing:
-                    Boolean(user?.organization) &&
-                    user?.organization?.status === OrganizationStatus.Active
-            },
-            {
-                label: t('general.feedback.organizationHasToHaveExchangeDeposit'),
-                passing: Boolean(exchangeAddress)
-            }
-        ]
-    };
-
-    canCreateDevice.value = canCreateDevice.rules.every((r) => r.passing);
-
-    return {
-        canCreateDevice
-    };
 }
 
 export const deviceById = (
